Fix authorization case
Paypal returns in separate authorization request no real state and returns "pending" with reason code "authorization".

So I add a check to handle this and skip throwing such a authorization pending exception.

From the docs:
authorization – The payment is pending because it has been authorized but not settled. You must capture the funds first.